requirements.txt
aiohttp>=3.8.0
asyncio
python-socketio